feat: add buttons v1

+++ bot/models/dto.go
@@ -6,13 +6,13 @@ type CreateEventRequest struct {
 	Emails      []string `json:"recipients_emails"`
 	StartDate   string   `json:"start_datetime"`
 	EndDate     string   `json:"end_datetime"`
-	Timezone    int64    `json:"timezone"`
+	Timezone    int      `json:"timezone"`
 }
 
 type LLMMessageRequest struct {
 	Messages []string `json:"messages"`
 	Language string   `json:"language"`
-	Timezone int64    `json:"timezone"`
+	Timezone int      `json:"timezone"`
 }
 
 type LLMResponse struct {
